test(db): move covers + settings + metadata_overrides tests
