+++ db/src/queries_tests.rs
@@ -37,162 +37,6 @@ mod tests {
     use sqlx::{Row, SqlitePool};
 
     #[tokio::test]
-    async fn cover_returns_none_when_file_missing() {
-        let _covers = CoversTempDir::new("missing");
-        let pool = init_db("sqlite::memory:").await.unwrap();
-        replace_books(
-            &pool,
-            "/lib",
-            vec![indexed(
-                "a.epub",
-                Some("A"),
-                &["A"],
-                &[],
-                None,
-                Some(("image/jpeg", b"BYTES")),
-            )],
-        )
-        .await
-        .unwrap();
-        let books = list_books(&pool, "/lib").await.unwrap();
-        let uuid: String = sqlx::query_scalar("SELECT uuid FROM books WHERE id = ?")
-            .bind(books[0].id)
-            .fetch_one(&pool)
-            .await
-            .unwrap();
-        // Remove the file out from under the DB — get_cover should report
-        // None, not error.
-        let _ = std::fs::remove_file(cover_path_for(&uuid, "jpg"));
-        assert!(get_cover(&pool, books[0].id).await.unwrap().is_none());
-    }
-
-    /// F5.1 / #107: when a `metadata_overrides` row sets
-    /// `has_cover_override = 1` and an `override-<uuid>.<ext>` file exists
-    /// on disk, `get_cover` returns the override bytes — not the scanned
-    /// cover. Single-query form must preserve this precedence.
-    #[tokio::test]
-    async fn cover_returns_override_when_flag_set() {
-        let _covers = CoversTempDir::new("override_set");
-        let pool = init_db("sqlite::memory:").await.unwrap();
-        let user_id = crate::auth::create_user(&pool, "admin", "securepassword1")
-            .await
-            .unwrap()
-            .id;
-
-        replace_books(
-            &pool,
-            "/lib",
-            vec![indexed(
-                "a.epub",
-                Some("A"),
-                &["A"],
-                &[],
-                None,
-                Some(("image/jpeg", b"ORIGINAL")),
-            )],
-        )
-        .await
-        .unwrap();
-
-        let books = list_books(&pool, "/lib").await.unwrap();
-        let uuid = books[0].unique_identifier.clone().unwrap();
-
-        // Mark cover-override + drop the override file on disk.
-        write_override_cover(&uuid, "image/png", b"OVERRIDE").unwrap();
-        upsert_metadata_overrides(&pool, &uuid, &MetadataOverrides::default(), true, user_id)
-            .await
-            .unwrap();
-
-        let cover = get_cover(&pool, books[0].id).await.unwrap();
-        assert_eq!(cover, Some(("image/png".into(), b"OVERRIDE".to_vec())));
-    }
-
-    /// F5.1 / #107: with no `metadata_overrides` row, `get_cover` falls
-    /// through to the scanned `<uuid>.<ext>` cover. The LEFT JOIN must
-    /// not filter the book out when no override row exists.
-    #[tokio::test]
-    async fn cover_returns_original_when_no_override_row() {
-        let _covers = CoversTempDir::new("override_absent");
-        let pool = init_db("sqlite::memory:").await.unwrap();
-
-        replace_books(
-            &pool,
-            "/lib",
-            vec![indexed(
-                "a.epub",
-                Some("A"),
-                &["A"],
-                &[],
-                None,
-                Some(("image/jpeg", b"ORIGINAL")),
-            )],
-        )
-        .await
-        .unwrap();
-
-        let books = list_books(&pool, "/lib").await.unwrap();
-        let cover = get_cover(&pool, books[0].id).await.unwrap();
-        assert_eq!(cover, Some(("image/jpeg".into(), b"ORIGINAL".to_vec())));
-    }
-
-    /// F5.1 / #107: a `metadata_overrides` row with
-    /// `has_cover_override = 0` (text-only edits, no cover swap) must
-    /// resolve to the scanned cover, not the override path.
-    #[tokio::test]
-    async fn cover_returns_original_when_override_flag_unset() {
-        let _covers = CoversTempDir::new("override_flag_off");
-        let pool = init_db("sqlite::memory:").await.unwrap();
-        let user_id = crate::auth::create_user(&pool, "admin", "securepassword1")
-            .await
-            .unwrap()
-            .id;
-
-        replace_books(
-            &pool,
-            "/lib",
-            vec![indexed(
-                "a.epub",
-                Some("A"),
-                &["A"],
-                &[],
-                None,
-                Some(("image/jpeg", b"ORIGINAL")),
-            )],
-        )
-        .await
-        .unwrap();
-
-        let books = list_books(&pool, "/lib").await.unwrap();
-        let uuid = books[0].unique_identifier.clone().unwrap();
-
-        // Override row exists with text edits but no cover swap.
-        upsert_metadata_overrides(
-            &pool,
-            &uuid,
-            &MetadataOverrides {
-                title: Some("Edited".into()),
-                ..Default::default()
-            },
-            false,
-            user_id,
-        )
-        .await
-        .unwrap();
-
-        let cover = get_cover(&pool, books[0].id).await.unwrap();
-        assert_eq!(cover, Some(("image/jpeg".into(), b"ORIGINAL".to_vec())));
-    }
-
-    /// #107: `get_cover` for a non-existent book id returns `Ok(None)`
-    /// (not an error). The LEFT JOIN must not change this contract.
-    #[tokio::test]
-    async fn cover_returns_none_for_missing_book_id() {
-        let _covers = CoversTempDir::new("missing_book");
-        let pool = init_db("sqlite::memory:").await.unwrap();
-        assert!(get_cover(&pool, 999_999).await.unwrap().is_none());
-    }
-
-    #[tokio::test]
     async fn library_from_db_returns_empty_for_none_path() {
         let pool = init_db("sqlite::memory:").await.unwrap();
         let lib = library_from_db(&pool, None).await.unwrap();
@@ -739,229 +583,6 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn set_settings_prunes_library_when_ebook_path_changes() {
-        let _covers = CoversTempDir::new("prune-change");
-        let pool = init_db("sqlite::memory:").await.unwrap();
-        set_settings(
-            &pool,
-            &Settings {
-                ebook_library_path: Some("/old".into()),
-                audiobook_library_path: None,
-            },
-        )
-        .await
-        .unwrap();
-        replace_books(
-            &pool,
-            "/old",
-            vec![indexed(
-                "a.epub",
-                Some("Dracula"),
-                &["Stoker"],
-                &[],
-                None,
-                None,
-            )],
-        )
-        .await
-        .unwrap();
-        assert_eq!(list_books(&pool, "/old").await.unwrap().len(), 1);
-
-        set_settings(
-            &pool,
-            &Settings {
-                ebook_library_path: Some("/new".into()),
-                audiobook_library_path: None,
-            },
-        )
-        .await
-        .unwrap();
-
-        assert!(list_books(&pool, "/old").await.unwrap().is_empty());
-        let library_count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM libraries")
-            .fetch_one(&pool)
-            .await
-            .unwrap();
-        assert_eq!(library_count, 0);
-        let book_count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM books")
-            .fetch_one(&pool)
-            .await
-            .unwrap();
-        assert_eq!(book_count, 0);
-        let fts_count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM books_fts")
-            .fetch_one(&pool)
-            .await
-            .unwrap();
-        assert_eq!(fts_count, 0);
-    }
-
-    #[tokio::test]
-    async fn set_settings_keeps_libraries_still_configured() {
-        let _covers = CoversTempDir::new("prune-keep");
-        let pool = init_db("sqlite::memory:").await.unwrap();
-        set_settings(
-            &pool,
-            &Settings {
-                ebook_library_path: Some("/books".into()),
-                audiobook_library_path: Some("/audio".into()),
-            },
-        )
-        .await
-        .unwrap();
-        replace_books(
-            &pool,
-            "/books",
-            vec![indexed("a.epub", Some("A"), &["X"], &[], None, None)],
-        )
-        .await
-        .unwrap();
-
-        set_settings(
-            &pool,
-            &Settings {
-                ebook_library_path: Some("/books".into()),
-                audiobook_library_path: Some("/audio".into()),
-            },
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(list_books(&pool, "/books").await.unwrap().len(), 1);
-    }
-
-    #[tokio::test]
-    async fn set_settings_none_removes_library_data() {
-        let _covers = CoversTempDir::new("prune-clear");
-        let pool = init_db("sqlite::memory:").await.unwrap();
-        set_settings(
-            &pool,
-            &Settings {
-                ebook_library_path: Some("/books".into()),
-                audiobook_library_path: None,
-            },
-        )
-        .await
-        .unwrap();
-        replace_books(
-            &pool,
-            "/books",
-            vec![indexed("a.epub", Some("A"), &["X"], &[], None, None)],
-        )
-        .await
-        .unwrap();
-
-        set_settings(
-            &pool,
-            &Settings {
-                ebook_library_path: None,
-                audiobook_library_path: None,
-            },
-        )
-        .await
-        .unwrap();
-
-        let library_count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM libraries")
-            .fetch_one(&pool)
-            .await
-            .unwrap();
-        assert_eq!(library_count, 0);
-    }
-
-    /// Exercises the batched (#149) `prune_orphan_libraries` path across more
-    /// than one chunk. The IN-list is chunked at 500 ids to stay under
-    /// SQLite's bind-parameter cap, so seeding more orphaned libraries than a
-    /// single chunk holds is what actually verifies the chunking loop iterates
-    /// (a regression that dropped or mis-bound a later chunk would otherwise go
-    /// undetected). Seeds 1001 libraries — three chunks of 500 / 500 / 1 — each
-    /// with a book, then prunes them all in one transaction and asserts every
-    /// cover UUID is collected (so the lookup is verified to span all chunks)
-    /// and every row is gone. Only a handful of rows get an on-disk cover:
-    /// materializing 1001 files would be slow, and cover deletion is exercised
-    /// by the tracked subset, which straddles the chunk boundaries.
-    #[tokio::test]
-    async fn prune_orphan_libraries_batches_across_many_libraries() {
-        // > 2 full chunks of 500 → the chunk loop runs three times.
-        const LIBRARY_COUNT: usize = 1001;
-        // Indices spanning every chunk: first row, the first row of the second
-        // chunk, a mid-chunk row, and the final (third-chunk) row.
-        const MATERIALIZED_COVER_INDICES: [usize; 4] = [0, 500, 750, LIBRARY_COUNT - 1];
-
-        let _covers = CoversTempDir::new("prune-batch");
-        let pool = init_db("sqlite::memory:").await.unwrap();
-
-        // Seed the orphaned libraries directly. `keep = []` below marks all of
-        // them as orphans regardless of path.
-        let mut expected_uuids: Vec<String> = Vec::with_capacity(LIBRARY_COUNT);
-        let mut materialized_uuids: Vec<String> = Vec::new();
-        for i in 0..LIBRARY_COUNT {
-            let path = format!("/orphan-{i}");
-            let library_id: i64 = sqlx::query_scalar(
-                "INSERT INTO libraries (path, display_name) VALUES (?, ?) RETURNING id",
-            )
-            .bind(&path)
-            .bind(format!("Orphan {i}"))
-            .fetch_one(&pool)
-            .await
-            .unwrap();
-
-            let uuid = format!("uuid-{i}");
-            sqlx::query(
-                "INSERT INTO books (uuid, library_id, path, title, has_cover)
-                 VALUES (?, ?, ?, ?, 1)",
-            )
-            .bind(&uuid)
-            .bind(library_id)
-            .bind(format!("{path}/book.epub"))
-            .bind(format!("Book {i}"))
-            .execute(&pool)
-            .await
-            .unwrap();
-
-            // Materialize an on-disk cover for a few rows spanning the chunk
-            // boundaries so deletion is observable without writing 1001 files.
-            if MATERIALIZED_COVER_INDICES.contains(&i) {
-                write_cover_file(&uuid, "image/jpeg", b"fake-jpeg").unwrap();
-                assert!(cover_path_for(&uuid, "jpg").exists());
-                materialized_uuids.push(uuid.clone());
-            }
-
-            expected_uuids.push(uuid);
-        }
-
-        let mut tx = pool.begin().await.unwrap();
-        let mut orphan_uuids = prune_orphan_libraries(&mut tx, &[]).await.unwrap();
-        tx.commit().await.unwrap();
-
-        orphan_uuids.sort();
-        expected_uuids.sort();
-        assert_eq!(
-            orphan_uuids, expected_uuids,
-            "every orphaned book's cover UUID should be collected across all chunks"
-        );
-
-        let library_count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM libraries")
-            .fetch_one(&pool)
-            .await
-            .unwrap();
-        assert_eq!(library_count, 0);
-        let book_count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM books")
-            .fetch_one(&pool)
-            .await
-            .unwrap();
-        assert_eq!(book_count, 0);
-
-        // The caller deletes cover files post-commit; verify the collected
-        // UUIDs drive removal of every materialized cover (one per chunk).
-        delete_cover_files_for(&orphan_uuids);
-        for uuid in &materialized_uuids {
-            assert!(
-                !cover_path_for(uuid, "jpg").exists(),
-                "cover for {uuid} should be deleted"
-            );
-        }
-    }
-
-    #[tokio::test]
     async fn get_book_returns_none_for_missing_id() {
         let pool = init_db("sqlite::memory:").await.unwrap();
         let result = get_book(&pool, 9999).await.unwrap();
@@ -1505,303 +1126,6 @@ mod tests {
         );
     }
 
-    // -----------------------------------------------------------------
-    // F5.1 Metadata overrides
-    // -----------------------------------------------------------------
-
-    #[tokio::test]
-    async fn upsert_and_get_metadata_overrides_roundtrips() {
-        let pool = init_db("sqlite::memory:").await.unwrap();
-        // Create a user for updated_by.
-        let user_id = crate::auth::create_user(&pool, "admin", "securepassword1")
-            .await
-            .unwrap()
-            .id;
-
-        let ov = MetadataOverrides {
-            title: Some("New Title".into()),
-            description: Some("A new description".into()),
-            ..Default::default()
-        };
-        upsert_metadata_overrides(&pool, "test-uuid-1", &ov, false, user_id)
-            .await
-            .unwrap();
-
-        let (loaded, has_cover) = get_metadata_overrides(&pool, "test-uuid-1")
-            .await
-            .unwrap()
-            .expect("overrides should exist");
-        assert_eq!(loaded.title, Some("New Title".into()));
-        assert_eq!(loaded.description, Some("A new description".into()));
-        assert_eq!(loaded.publisher, None);
-        assert!(!has_cover);
-    }
-
-    #[tokio::test]
-    async fn merge_metadata_overrides_accumulates_fields_and_preserves_cover_flag() {
-        let pool = init_db("sqlite::memory:").await.unwrap();
-        let user_id = crate::auth::create_user(&pool, "admin", "securepassword1")
-            .await
-            .unwrap()
-            .id;
-
-        // Seed an existing override carrying a title AND a user-uploaded cover.
-        let initial = MetadataOverrides {
-            title: Some("First Title".into()),
-            ..Default::default()
-        };
-        upsert_metadata_overrides(&pool, "merge-uuid", &initial, true, user_id)
-            .await
-            .unwrap();
-
-        // A later edit touching only `description` must not clobber the title
-        // (the incremental-edit contract the TOCTOU race nullified) and must
-        // not reset the cover flag (the pre-#166 reset bug).
-        let edit = MetadataOverrides {
-            description: Some("Added later".into()),
-            ..Default::default()
-        };
-        merge_metadata_overrides(&pool, "merge-uuid", &edit, user_id)
-            .await
-            .unwrap();
-
-        let (loaded, has_cover) = get_metadata_overrides(&pool, "merge-uuid")
-            .await
-            .unwrap()
-            .expect("overrides should exist");
-        assert_eq!(
-            loaded.title,
-            Some("First Title".into()),
-            "prior title must survive a description-only merge"
-        );
-        assert_eq!(loaded.description, Some("Added later".into()));
-        assert!(
-            has_cover,
-            "has_cover_override must carry forward across a text-only merge"
-        );
-    }
-
-    #[tokio::test]
-    async fn merge_metadata_overrides_creates_row_when_absent() {
-        let pool = init_db("sqlite::memory:").await.unwrap();
-        let user_id = crate::auth::create_user(&pool, "admin", "securepassword1")
-            .await
-            .unwrap()
-            .id;
-        let edit = MetadataOverrides {
-            title: Some("Fresh".into()),
-            ..Default::default()
-        };
-        merge_metadata_overrides(&pool, "fresh-uuid", &edit, user_id)
-            .await
-            .unwrap();
-        let (loaded, has_cover) = get_metadata_overrides(&pool, "fresh-uuid")
-            .await
-            .unwrap()
-            .expect("overrides should exist");
-        assert_eq!(loaded.title, Some("Fresh".into()));
-        assert!(!has_cover, "a brand-new merged row has no cover override");
-    }
-
-    #[tokio::test]
-    async fn get_metadata_overrides_returns_none_when_absent() {
-        let pool = init_db("sqlite::memory:").await.unwrap();
-        let result = get_metadata_overrides(&pool, "nonexistent-uuid")
-            .await
-            .unwrap();
-        assert!(result.is_none());
-    }
-
-    #[tokio::test]
-    async fn delete_metadata_overrides_removes_row() {
-        let pool = init_db("sqlite::memory:").await.unwrap();
-        let user_id = crate::auth::create_user(&pool, "admin", "securepassword1")
-            .await
-            .unwrap()
-            .id;
-        let ov = MetadataOverrides {
-            title: Some("Override".into()),
-            ..Default::default()
-        };
-        upsert_metadata_overrides(&pool, "del-uuid", &ov, false, user_id)
-            .await
-            .unwrap();
-        assert!(get_metadata_overrides(&pool, "del-uuid")
-            .await
-            .unwrap()
-            .is_some());
-
-        delete_metadata_overrides(&pool, "del-uuid").await.unwrap();
-        assert!(get_metadata_overrides(&pool, "del-uuid")
-            .await
-            .unwrap()
-            .is_none());
-    }
-
-    /// Bug #1: saving a title override must rebuild `books_fts` so search
-    /// finds the new title and stops matching the original one.
-    #[tokio::test]
-    async fn upsert_metadata_overrides_rebuilds_fts_for_title() {
-        let _covers = CoversTempDir::new("fts_override_title");
-        let pool = init_db("sqlite::memory:").await.unwrap();
-        let user_id = crate::auth::create_user(&pool, "admin", "securepassword1")
-            .await
-            .unwrap()
-            .id;
-
-        replace_books(
-            &pool,
-            "/lib",
-            vec![indexed(
-                "a.epub",
-                Some("Original Title"),
-                &["Author A"],
-                &[],
-                None,
-                None,
-            )],
-        )
-        .await
-        .unwrap();
-
-        // Sanity: search finds the original title.
-        let hits = search_books(&pool, "/lib", "Original").await.unwrap();
-        assert_eq!(hits.len(), 1);
-
-        // Save an override that changes the title.
-        let uuid = list_books(&pool, "/lib").await.unwrap()[0]
-            .unique_identifier
-            .clone()
-            .unwrap();
-        let ov = MetadataOverrides {
-            title: Some("Brand New Title".into()),
-            ..Default::default()
-        };
-        upsert_metadata_overrides(&pool, &uuid, &ov, false, user_id)
-            .await
-            .unwrap();
-
-        // Search now matches the overridden title and no longer the original.
-        let new_hits = search_books(&pool, "/lib", "Brand").await.unwrap();
-        assert_eq!(new_hits.len(), 1);
-        assert_eq!(new_hits[0].title.as_deref(), Some("Brand New Title"));
-        let old_hits = search_books(&pool, "/lib", "Original").await.unwrap();
-        assert!(
-            old_hits.is_empty(),
-            "FTS still matches the pre-override title"
-        );
-    }
-
-    /// Bug #1: the palette uses the same `books_fts` table, so the override
-    /// rebuild must also surface there.
-    #[tokio::test]
-    async fn upsert_metadata_overrides_rebuilds_fts_for_palette() {
-        let _covers = CoversTempDir::new("fts_override_palette");
-        let pool = init_db("sqlite::memory:").await.unwrap();
-        let user_id = crate::auth::create_user(&pool, "admin", "securepassword1")
-            .await
-            .unwrap()
-            .id;
-
-        replace_books(
-            &pool,
-            "/lib",
-            vec![indexed(
-                "p.epub",
-                Some("Scanned Title"),
-                &["Author"],
-                &[],
-                None,
-                None,
-            )],
-        )
-        .await
-        .unwrap();
-
-        let uuid = list_books(&pool, "/lib").await.unwrap()[0]
-            .unique_identifier
-            .clone()
-            .unwrap();
-        let ov = MetadataOverrides {
-            title: Some("Edited Palette Title".into()),
-            ..Default::default()
-        };
-        upsert_metadata_overrides(&pool, &uuid, &ov, false, user_id)
-            .await
-            .unwrap();
-
-        let palette = search_palette(&pool, "/lib", "Edited").await.unwrap();
-        assert_eq!(palette.books.len(), 1);
-    }
-
-    /// Bug #1 follow-on: deleting the override should restore the FTS row
-    /// to the canonical scanned values.
-    #[tokio::test]
-    async fn delete_metadata_overrides_restores_fts() {
-        let _covers = CoversTempDir::new("fts_override_revert");
-        let pool = init_db("sqlite::memory:").await.unwrap();
-        let user_id = crate::auth::create_user(&pool, "admin", "securepassword1")
-            .await
-            .unwrap()
-            .id;
-
-        replace_books(
-            &pool,
-            "/lib",
-            vec![indexed(
-                "r.epub",
-                Some("Canonical Title"),
-                &["Author"],
-                &[],
-                None,
-                None,
-            )],
-        )
-        .await
-        .unwrap();
-
-        let uuid = list_books(&pool, "/lib").await.unwrap()[0]
-            .unique_identifier
-            .clone()
-            .unwrap();
-
-        upsert_metadata_overrides(
-            &pool,
-            &uuid,
-            &MetadataOverrides {
-                title: Some("Temporary Override".into()),
-                ..Default::default()
-            },
-            false,
-            user_id,
-        )
-        .await
-        .unwrap();
-        assert_eq!(
-            search_books(&pool, "/lib", "Temporary")
-                .await
-                .unwrap()
-                .len(),
-            1
-        );
-
-        delete_metadata_overrides(&pool, &uuid).await.unwrap();
-
-        // FTS is back to the canonical title; the override token no longer
-        // matches.
-        assert_eq!(
-            search_books(&pool, "/lib", "Canonical")
-                .await
-                .unwrap()
-                .len(),
-            1
-        );
-        assert!(search_books(&pool, "/lib", "Temporary")
-            .await
-            .unwrap()
-            .is_empty());
-    }
-
     #[tokio::test]
     async fn get_book_merges_scalar_overrides() {
         let _covers = CoversTempDir::new("merge_scalar");
@@ -2262,82 +1586,6 @@ mod tests {
         assert!(!b.has_override);
         assert_eq!(c.title.as_deref(), Some("Edited C"));
         assert!(c.has_override);
-    }
-
-    #[tokio::test]
-    async fn delete_overrides_reverts_to_scanned() {
-        let _covers = CoversTempDir::new("revert");
-        let pool = init_db("sqlite::memory:").await.unwrap();
-        let user_id = crate::auth::create_user(&pool, "admin", "securepassword1")
-            .await
-            .unwrap()
-            .id;
-
-        replace_books(
-            &pool,
-            "/lib",
-            vec![indexed(
-                "revert.epub",
-                Some("Original"),
-                &["Author"],
-                &["fiction"],
-                None,
-                None,
-            )],
-        )
-        .await
-        .unwrap();
-
-        let books = list_books(&pool, "/lib").await.unwrap();
-        let uuid = books[0].unique_identifier.clone().unwrap();
-        let id = books[0].id;
-
-        // Override.
-        upsert_metadata_overrides(
-            &pool,
-            &uuid,
-            &MetadataOverrides {
-                title: Some("Changed".into()),
-                subjects: Some(vec!["sci-fi".into()]),
-                ..Default::default()
-            },
-            false,
-            user_id,
-        )
-        .await
-        .unwrap();
-        let merged = get_book(&pool, id).await.unwrap().unwrap();
-        assert_eq!(merged.title.as_deref(), Some("Changed"));
-
-        // Delete overrides — should revert to scanned.
-        delete_metadata_overrides(&pool, &uuid).await.unwrap();
-        let reverted = get_book(&pool, id).await.unwrap().unwrap();
-        assert_eq!(reverted.title.as_deref(), Some("Original"));
-        assert_eq!(reverted.subjects, vec!["fiction"]);
-        assert!(!reverted.has_override);
-    }
-
-    /// Verify that `MetadataOverrides::merge` correctly layers a second edit
-    /// on top of a first without losing the first edit's fields.
-    #[tokio::test]
-    async fn merge_preserves_prior_overrides() {
-        let first = MetadataOverrides {
-            title: Some("Edited Title".into()),
-            publisher: Some("Edited Publisher".into()),
-            ..Default::default()
-        };
-        let second = MetadataOverrides {
-            description: Some("New description".into()),
-            ..Default::default()
-        };
-        let merged = first.merge(&second);
-        // second's description wins
-        assert_eq!(merged.description.as_deref(), Some("New description"));
-        // first's title and publisher are preserved (not wiped by None)
-        assert_eq!(merged.title.as_deref(), Some("Edited Title"));
-        assert_eq!(merged.publisher.as_deref(), Some("Edited Publisher"));
-        // unset in both stays None
-        assert_eq!(merged.language, None);
     }
 
     // ── search_palette ──────────────────────────────────────────────
